+++ packages/flutter_adaptive_cards_fs/lib/flutter_adaptive_cards_fs.dart
@@ -16,6 +16,7 @@ export 'package:flutter_adaptive_cards_fs/src/models/data_query.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/fact.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/media_source.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/refresh_config.dart';
+export 'package:flutter_adaptive_cards_fs/src/models/text_run.dart';
 
 /// needed so we can create the registry to pass into the AdaptiveCard constructor
 export 'package:flutter_adaptive_cards_fs/src/registry.dart';